Move level proxy modules to "world" package
Introduce new interface for client level proxy, rename existing
type to cloned client level to match purpose. Allow other types
to conform to client level proxy type.
