+++ common/src/main/java/com/sonicether/soundphysics/world/ClientLevelProxy.java
@@ -1,0 +1,5 @@
+package com.sonicether.soundphysics.world;
+
+import net.minecraft.world.level.BlockGetter;
+
+public interface ClientLevelProxy extends BlockGetter {};
