+++ common/src/main/java/com/sonicether/soundphysics/world/ClonedLevelHeightAccessor.java
@@ -1,4 +1,4 @@
-package com.sonicether.soundphysics.models;
+package com.sonicether.soundphysics.world;
 
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelHeightAccessor;
